no message
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@787 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ trunk/rdo_lib/rdo_parser/rdosmr.cpp
@@ -63,7 +63,7 @@ void RDOSMR::setFrameNumber( int value, const YYLTYPE& pos )
 	if ( value <= 0 ) {
 		parser()->error( pos, "Номер кадра должен быть больше нуля" );
 	}
-	if ( parser()->runtime()->allFrames.size() + 1 <= value ) {
+	if ( parser()->runtime()->allFrames.size() + 1 <= (unsigned int)value ) {
 		parser()->error( pos, rdo::format("Несуществующий кадр: %d", value) );
 	}
 	frameNumber = value;
